(GH-16) Fix broken interpolated string highlighting
Previously interpolated text within strings would not highlight correctly on
various combinations of valid and invalid variable names and function calls, for
example names starting with underscore.  This commit updates the variable regex
matchers to align with the Puppet Language and ensures that the highlighting
looks correct.

Also note that the use of the contentName = source.puppet token.  This is used
to "reset" the token list for editors so the colouring can be set correctly.
This commit also adds tests for valid variable names within different sections
of an interpolated string.