(GH-22) Fix tokenizing of heredoc strings
Previously the syntax did not understand heredoc strings.  This commit updates
the language to parse bare and interpolated heredoc strings as per [1].

Note however that this does not tokenize multiple heredocs on the same line or
using heredocs as string types which can then have functions on the right hand
side. [2] [3]

[1] https://github.com/puppetlabs/puppet-specifications/blob/master/language/heredoc.md
[2] https://github.com/puppetlabs/puppet-specifications/blob/master/language/heredoc.md#multiple-heredoc-on-the-same-line
[3] https://github.com/puppetlabs/puppet-specifications/blob/master/language/heredoc.md#additional-semantics