(maint) Make regex matching lazy for syntax highlighting
The regex used to match regexes in puppet uses (.+) to match one or more of
anything in between two slashes. However, if there is another slash anywhere
on the line after the regex the (.+) will match between the first slash of the
regex and the other slash that's not part of the regex.

An example:

/puppet-agent-5..*/ => 'puppet5/',

In that line the syntax highlighter will match everything between the first
slash of the regex and the slash in the string as one regex.

This commit changes the match to the following:

1: The (.+) matcher is made to be 'lazy' so that it matches the smallest
string possible.

2: The final matcher (?:\/) is updated to include the negated set including
the backslash so that the final matcher matches a final slash with a
preceeding token that's anything _except_ \.

This should make the regex match the smallest possible string that begins with
/ and ends with / but does not end with \/.