(maint) Update vscode-textmate module to 4.x
VSCode now uses the vscode-textmate module of 4.x which had breaking API
changes.  This commit updates the module version to 4.x and updates the tests
and inspect script to use the new API surface.

Note that no tests or actual Textmate language files were modified as part of
this process.