(GH-38) Tokenise Function and Defined Type parameters
Previously function parameters were not being tokenised correctly as per the
newer Puppet 4 API Data types.

This commit updates the function and defined type parameters to be similar to
that of classes and plans, whereby Puppet Data types etc. are tokenised
correctly.  This commit also adds tests for this scenario, and adds negative
tests to ensure that invalid names are not tokenised.