(GH-38)(GH-39) Fix tokenising of class, plan and node names
Previously the node, plan and class names were far too loose in what
characters were allowed:

* Plans and classes should only allow qualified names, with or without name
  segments.  Previously they allowed quotes, hyphens and capital letters
* Nodes can only accept strings for their names, with any characters, and have
  no parameters.  Previously nodes shared the class definition which is very
  different.

This commit moves the node detection to it's own section and tightens up the
allowed names for classes and plans.  This commit also adds tests for allowed
and disallowed names for plans, classes and nodes.