(GH-43) Fix syntax highlighting for resource references and chain arrows
Previously the sytnax highlighting for resource references did not work for
valid names.  Although Puppet Types do not allow an underscore, Resource
References, which are technically types, do.  This commit changes the regex to
allow underscores as there's no way for the parser to know what kind of
reference it actually is

Previously the highlighting of class names after a chain arrow was failing. This
was due to the chain arrow not absorbing the whitespace after it.  This commit
changes the chain arrow regex to consume trailing and leading whitespace.

This commit also adds tests for these scenarios.