(GH-65) Tokenise Node names correctly
Previously the tokeniser always thought node names must a be a single string,
however they can also be regular expressions and the keyword default. And they
can be a list of these separate by commas.  This commuit updates the node
regular expressions for all of these types of node names and adds tests for
these scenarios